Add files via upload
Version 2 " This mod replaces traders from factions with a higher tech level than the player's faction. The tech level difference can be adjusted in the mod settings. By default, only factions with a tech level up to one level higher than the player's faction are allowed. The list of allowed factions is updated whenever the player's tech level changes or a new game is started or loaded."